chore(conversion): swap hub/spoke conversion for paas
